Update multi_chain_comparison.py (#2025)
Encountered a type error locally since my signature was returning a float (score), which doesn't have the `strip()` method. Once we cast this answer to a string, things work smoothly. 

NOTE: It seems like the answers provided within the completions only contain the final output field, but this does not generalize to cases where the signature has multiple output fields.