Update chain_of_thought.py (#8006)
* Update chain_of_thought.py

Allow use of custom CoT representations (e.g., List of strings) instead of forcing CoT be a string output. 
Previously, even if the user passed in a custom definition for the reasoning field, the type of the reasoning field output would be a string. 
This change ensures that the type of the field is consistent with its annotation, and allows for users to specify a custom type for reasoning without creating a custom field for it. 
Also, this change introduces a docstring and type hints into the ChainOfThought module.

* Update chain_of_thought.py

Fixed error due to missing import.

* Update chain_of_thought.py

Avoid circular import and ensure there are two newlines between imports and initial code.

* Update chain_of_thought.py

Avoided circular import caused by referencing `dspy.OutputField`